Added searching to users management

+++ HealthCenter/src/main/java/com/github/marcinseweryn/service/UserServiceImpl.java
@@ -16,15 +16,18 @@ public class UserServiceImpl implements UserService{
 
 	@Autowired
 	private UserDAO userDAO;
+	
+	DateFormat df = new SimpleDateFormat("yyyy-MM-dd");	
 
 	public void addUser(User user){	
-		
-		BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-		String hashedPassword = passwordEncoder.encode(user.getPassword());
-		
-		user.setPassword(hashedPassword);
-		
-		userDAO.addUser(user);
+
+		if(!user.getPesel().equals("")){
+			BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+			String hashedPassword = passwordEncoder.encode(user.getPassword());
+			
+			user.setPassword(hashedPassword);
+			userDAO.addUser(user);
+		}
 	}
 
 
@@ -34,58 +37,112 @@ public class UserServiceImpl implements UserService{
 
 
 	public void deleteUsers(List<Integer> usersIDs) {
-		userDAO.deleteUsers(usersIDs);	
+
+		if(usersIDs.size() != 0){
+			userDAO.deleteUsers(usersIDs);	
+		}
 	}
 
 
 	@Override
-	public void updateUsers(List<Integer> usersIDs, User user) {
-		DateFormat df = new SimpleDateFormat("yyyy-MM-dd");		
+	public void updateUsers(List<Integer> usersIDs, User user) {	
 		String columns = "";
 		
 		if(!user.getPesel().equals("")){
-			columns += "pesel = '"+user.getPesel()+"' ,"; 
+			columns += "pesel = '" + user.getPesel() + "' ,"; 
 		}
 		if(!user.getGender().equals("")){
-			columns += "gender = '"+user.getGender()+"' ,";
+			columns += "gender = '" + user.getGender() + "' ,";
 		}
 		if(!user.getName().equals("")){
-			columns += "name = '"+user.getName()+"' ,";
+			columns += "name = '" + user.getName() + "' ,";
 		}
 		if(!user.getSurname().equals("")){
-			columns += "surname = '"+user.getSurname()+"' ,";
+			columns += "surname = '" + user.getSurname() + "' ,";
 		}
 		if(user.getBirthDate() != null){
-			columns += "birthDate = '"+df.format(user.getBirthDate())+"' ,";
+			columns += "birthDate = '" + df.format(user.getBirthDate()) + "' ,";
 		}
 		if(!user.getPassword().equals("")){
 			BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
 			String hashedPassword = passwordEncoder.encode(user.getPassword());
-			columns += "password = '"+hashedPassword+"' ,";
+			columns += "password = '" + hashedPassword + "' ,";
 		}
 		if(!user.getStreetAddress().equals("")){
-			columns += "streetAddress = '"+user.getStreetAddress()+"' ,";
+			columns += "streetAddress = '" + user.getStreetAddress() + "' ,";
 		}
 		if(!user.getCity().equals("")){
-			columns += "city = '"+user.getCity()+"' ,";
+			columns += "city = '" + user.getCity() + "' ,";
 		}
 		if(!user.getPostalCode().equals("")){
-			columns += "postalCode = '"+user.getPostalCode()+"' ,";
+			columns += "postalCode = '" + user.getPostalCode() + "' ,";
 		}
 		if(!user.getPhone().equals("")){
-			columns += "phone = '"+user.getPhone()+"' ,";
+			columns += "phone = '" + user.getPhone() + "' ,";
 		}
 		if(!user.getEmail().equals("")){
-			columns += "email = '"+user.getEmail()+"' ,";
+			columns += "email = '" + user.getEmail() + "' ,";
 		}
 		if(!user.getRole().equals("")){
-			columns += "role = '"+user.getRole()+"' ,";
+			columns += "role = '" + user.getRole() + "' ,";
 		}
 		System.out.println(columns);
 		if(columns.length() > 0){
 			columns = columns.substring(0, columns.length() - 1);
 		}
+		
+		if(!columns.equals("")){
+			userDAO.updateUsers(usersIDs, user, columns);
+		}
+	}
+
+
+	@Override
+	public List<User> findUsers(User user) {
+		String columns = "";
+		
+		if(!user.getPesel().equals("")){
+			columns += "pesel = '" + user.getPesel() + "' and "; 
+		}
+		if(!user.getGender().equals("")){
+			columns += "gender = '" + user.getGender() + "' and ";
+		}
+		if(!user.getName().equals("")){
+			columns += "name = '" + user.getName() + "' and ";
+		}
+		if(!user.getSurname().equals("")){
+			columns += "surname = '" + user.getSurname() + "' and ";
+		}
+		if(user.getBirthDate() != null){
+			columns += "birthDate = '" + df.format(user.getBirthDate()) + "' and ";
+		}
+		if(!user.getStreetAddress().equals("")){
+			columns += "streetAddress = '" + user.getStreetAddress() + "' and ";
+		}
+		if(!user.getCity().equals("")){
+			columns += "city = '" + user.getCity() + "' and ";
+		}
+		if(!user.getPostalCode().equals("")){
+			columns += "postalCode = '" + user.getPostalCode() + "' and ";
+		}
+		if(!user.getPhone().equals("")){
+			columns += "phone = '" + user.getPhone() + "' and ";
+		}
+		if(!user.getEmail().equals("")){
+			columns += "email = '" + user.getEmail() + "' and ";
+		}
+		if(!user.getRole().equals("")){
+			columns += "role = '" + user.getRole() + "' and ";
+		}
 		System.out.println(columns);
-		userDAO.updateUsers(usersIDs, user, columns);
+		if(columns.length() > 0){
+			columns = columns.substring(0, columns.length() - 4);
+		}
+		
+		if(columns.equals("")){
+			return null;
+		}else{
+			return userDAO.findUsers(user, columns);
+		}		
 	}
 }
